delete legacy module
