+++ .idea/gradle.xml
@@ -30,7 +30,6 @@
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/samples" />
             <option value="$PROJECT_DIR$/feature/slide" />
-            <option value="$PROJECT_DIR$/legacy" />
             <option value="$PROJECT_DIR$/slide" />
           </set>
         </option>
