+++ app/src/main/java/net/kwmt27/jetpackcomposeplayground/NavHost.kt
@@ -57,19 +57,19 @@ import net.kwmt27.jetpackcomposeplayground.feature.samples.animation.UpdateTrans
 import net.kwmt27.jetpackcomposeplayground.feature.samples.coroutine.CancellationExceptionSample
 import net.kwmt27.jetpackcomposeplayground.feature.samples.image.CircleImageSample
 import net.kwmt27.jetpackcomposeplayground.feature.samples.livedata.SampleMutableLiveDataScreen
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.bottomsheet.BottomSheetLongDataSample
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.bottomsheet.BottomSheetSample
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.box.BoxSample
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.button.ButtonSample
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.edittext.OutlinedTextFieldSample
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.icon.UpIconSample
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.list.ListWithUpdatableItem
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.list.SampleHorizontalList
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.list.SampleVerticalList
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.list.StickyListSample
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.list.instagram.SampleInstagramSearchListLayout
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.state.ExpandableCardSample
-import net.kwmt27.jetpackcomposeplayground.legacy.feature.tablayout.TabLayoutSample
+import net.kwmt27.jetpackcomposeplayground.feature.samples.bottomsheet.BottomSheetLongDataSample
+import net.kwmt27.jetpackcomposeplayground.feature.samples.bottomsheet.BottomSheetSample
+import net.kwmt27.jetpackcomposeplayground.feature.samples.box.BoxSample
+import net.kwmt27.jetpackcomposeplayground.feature.samples.button.ButtonSample
+import net.kwmt27.jetpackcomposeplayground.feature.samples.edittext.OutlinedTextFieldSample
+import net.kwmt27.jetpackcomposeplayground.feature.samples.icon.UpIconSample
+import net.kwmt27.jetpackcomposeplayground.feature.samples.list.ListWithUpdatableItem
+import net.kwmt27.jetpackcomposeplayground.feature.samples.list.SampleHorizontalList
+import net.kwmt27.jetpackcomposeplayground.feature.samples.list.SampleVerticalList
+import net.kwmt27.jetpackcomposeplayground.feature.samples.list.StickyListSample
+import net.kwmt27.jetpackcomposeplayground.feature.samples.list.instagram.SampleInstagramSearchListLayout
+import net.kwmt27.jetpackcomposeplayground.feature.samples.state.ExpandableCardSample
+import net.kwmt27.jetpackcomposeplayground.feature.samples.tablayout.TabLayoutSample
 import net.kwmt27.jetpackcomposeplayground.slide.contents.potatotips82.PotatoTips82
 import net.kwmt27.jetpackcomposeplayground.slide.contents.shibuyaapk43.ShibuyaApk43Presentation
